Move factories and fixtures to their own modules
